+++ packages/swirl-components/src/components/swirl-inline-status/swirl-inline-status.tsx
@@ -30,10 +30,7 @@ export class SwirlInlineStatus {
   componentDidLoad() {
     this.forceIconProps(this.desktopMediaQuery.matches);
 
-    this.desktopMediaQuery.addEventListener?.(
-      "change",
-      this.desktopMediaQueryHandler
-    );
+    this.desktopMediaQuery.onchange = this.desktopMediaQueryHandler;
   }
 
   disconnectedCallback() {