+++ packages/swirl-components/src/components/swirl-resource-list/swirl-resource-list.tsx
@@ -1,20 +1,74 @@
-import { Component, Element, h, Host, Prop } from "@stencil/core";
+import {
+  Component,
+  Element,
+  Event,
+  EventEmitter,
+  h,
+  Host,
+  Prop,
+  State,
+  Watch,
+} from "@stencil/core";
+import Sortable, { SortableEvent } from "sortablejs";
 
 @Component({
-  shadow: true,
+  scoped: true,
+  shadow: false,
   styleUrl: "swirl-resource-list.css",
   tag: "swirl-resource-list",
 })
 export class SwirlResourceList {
   @Element() el: HTMLElement;
 
+  @Prop() allowDrag?: boolean;
+  @Prop() assistiveTextItemGrabbed?: string =
+    "Item grabbed. Use arrow keys to move item up or down. Use spacebar to save position.";
+  @Prop() assistiveTextItemMoving?: string = "Current position:";
+  @Prop() assistiveTextItemMoved?: string = "Item moved. New position:";
   @Prop() label?: string;
 
+  @State() assistiveText: string;
+
+  @Event() itemDrop: EventEmitter<{
+    item: HTMLSwirlResourceListItemElement;
+    oldIndex: number;
+    newIndex: number;
+  }>;
+
+  private dragging: HTMLSwirlResourceListItemElement;
+  private draggingStartIndex: number;
   private focusedIndex = 0;
+  private gridEl: HTMLElement;
   private items: HTMLSwirlResourceListItemElement[];
+  private observer: MutationObserver;
+  private sortable: Sortable;
 
   componentDidLoad() {
+    this.observeSlotChanges();
     this.collectItems();
+    this.setItemAllowDragState();
+    this.setupDragDrop();
+  }
+
+  disconnectedCallback() {
+    this.sortable?.destroy();
+    this.observer?.disconnect();
+  }
+
+  private observeSlotChanges() {
+    this.observer = new MutationObserver(() => {
+      this.collectItems();
+      this.setItemAllowDragState();
+      this.setupDragDrop();
+    });
+
+    this.observer.observe(this.el, { childList: true });
+  }
+
+  @Watch("allowDrag")
+  watchAllowDrag() {
+    this.setItemAllowDragState();
+    this.setupDragDrop();
   }
 
   private collectItems() {
@@ -28,15 +82,93 @@ export class SwirlResourceList {
     this.enableItemFocus(this.items[this.focusedIndex]);
   }
 
+  private getItemIndex(item: HTMLSwirlResourceListItemElement): number {
+    return this.items.map((i) => i).findIndex((i) => i === item);
+  }
+
   private removeItemsFromTabOrder() {
     this.items.forEach((item) =>
-      item.shadowRoot
+      item
         ?.querySelector(
           ".resource-list-item__content, .resource-list-file-item"
         )
         ?.setAttribute("tabIndex", "-1")
     );
   }
+
+  private setItemAllowDragState() {
+    if (this.allowDrag) {
+      this.items.forEach((item) => {
+        item.setAttribute("allow-drag", "true");
+        item.addEventListener("toggleDrag", this.toggleDrag);
+      });
+    } else {
+      this.items.forEach((item) => {
+        item.removeAttribute("allow-drag");
+        item.removeEventListener("toggleDrag", this.toggleDrag);
+      });
+    }
+  }
+
+  private setupDragDrop() {
+    if (Boolean(this.sortable)) {
+      this.sortable.destroy();
+    }
+
+    if (!this.allowDrag) {
+      return;
+    }
+
+    this.sortable = Sortable.create(this.gridEl, {
+      animation: 150,
+      draggable: "swirl-resource-list-item",
+      handle: ".resource-list-item__drag-handle",
+      onEnd: (event: SortableEvent) => {
+        this.itemDrop.emit({
+          item: event.item as HTMLSwirlResourceListItemElement,
+          oldIndex: event.oldDraggableIndex,
+          newIndex: event.newDraggableIndex,
+        });
+      },
+    });
+  }
+
+  private toggleDrag = (
+    event: CustomEvent<HTMLSwirlResourceListItemElement>
+  ) => {
+    const item = event.detail;
+
+    if (Boolean(this.dragging)) {
+      this.stopDrag(item);
+    } else {
+      this.startDrag(item);
+    }
+  };
+
+  private startDrag = (item: HTMLSwirlResourceListItemElement) => {
+    this.dragging = item;
+    this.draggingStartIndex = this.getItemIndex(this.dragging);
+
+    item.setAttribute("dragging", "true");
+
+    const index = this.getItemIndex(this.dragging);
+    this.focusItemAtIndex(index);
+
+    this.assistiveText = this.assistiveTextItemGrabbed;
+  };
+
+  private stopDrag = (item: HTMLSwirlResourceListItemElement) => {
+    const newIndex = this.getItemIndex(this.dragging);
+
+    this.dragging = undefined;
+    item.removeAttribute("dragging");
+
+    this.assistiveText = `${this.assistiveTextItemMoved} ${newIndex + 1}`;
+
+    this.itemDrop.emit({ item, oldIndex: this.draggingStartIndex, newIndex });
+
+    this.draggingStartIndex = undefined;
+  };
 
   private enableItemFocus(
     item?: HTMLSwirlResourceListItemElement,
@@ -46,7 +178,7 @@ export class SwirlResourceList {
       return;
     }
 
-    const interactiveElement = item.shadowRoot?.querySelector<HTMLElement>(
+    const interactiveElement = item.querySelector<HTMLElement>(
       ".resource-list-item__content, .resource-list-file-item"
     );
 
@@ -75,17 +207,70 @@ export class SwirlResourceList {
     this.focusedIndex = index;
   }
 
+  private moveDraggedItemDown() {
+    const nextSibling = this.dragging.nextElementSibling;
+
+    if (!Boolean(nextSibling)) {
+      return;
+    }
+
+    nextSibling.after(this.dragging);
+    this.collectItems();
+    this.focusItemAtIndex(this.getItemIndex(this.dragging));
+
+    this.assistiveText = `${this.assistiveTextItemMoving} ${
+      this.getItemIndex(this.dragging) + 1
+    }`;
+  }
+
+  private moveDraggedItemUp() {
+    const prevSibling = this.dragging.previousElementSibling;
+
+    if (!Boolean(prevSibling)) {
+      return;
+    }
+
+    prevSibling.before(this.dragging);
+    this.collectItems();
+    this.focusItemAtIndex(this.getItemIndex(this.dragging));
+
+    this.assistiveText = `${this.assistiveTextItemMoving} ${
+      this.getItemIndex(this.dragging) + 1
+    }`;
+  }
+
   private onKeyDown = (event: KeyboardEvent) => {
     if (event.key === "ArrowDown") {
       event.preventDefault();
-      this.focusItemAtIndex((this.focusedIndex + 1) % this.items.length);
+
+      if (!Boolean(this.dragging)) {
+        this.focusItemAtIndex((this.focusedIndex + 1) % this.items.length);
+      } else {
+        this.moveDraggedItemDown();
+      }
     } else if (event.key === "ArrowUp") {
       event.preventDefault();
 
-      const prevIndex =
-        this.focusedIndex === 0 ? this.items.length - 1 : this.focusedIndex - 1;
+      if (!Boolean(this.dragging)) {
+        const prevIndex =
+          this.focusedIndex === 0
+            ? this.items.length - 1
+            : this.focusedIndex - 1;
 
-      this.focusItemAtIndex(prevIndex);
+        this.focusItemAtIndex(prevIndex);
+      } else {
+        this.moveDraggedItemUp();
+      }
+    } else if (event.key === "Space" || event.key === "Enter") {
+      const target = event.composedPath()[0] as HTMLElement;
+
+      if (
+        Boolean(this.dragging) &&
+        !target?.classList.contains("resource-list-item__drag-handle")
+      ) {
+        event.preventDefault();
+        this.stopDrag(this.dragging);
+      }
     } else if (event.key === "Home") {
       event.preventDefault();
 
@@ -97,15 +282,19 @@ export class SwirlResourceList {
     }
   };
 
-  private onSlotChange = () => {
-    this.collectItems();
-  };
-
   render() {
     return (
       <Host onKeyDown={this.onKeyDown}>
-        <swirl-stack aria-label={this.label} role="grid">
-          <slot onSlotchange={this.onSlotChange}></slot>
+        <swirl-visually-hidden role="alert">
+          {this.assistiveText}
+        </swirl-visually-hidden>
+        <swirl-stack
+          aria-label={this.label}
+          class="resource-list"
+          ref={(el) => (this.gridEl = el)}
+          role="grid"
+        >
+          <slot></slot>
         </swirl-stack>
       </Host>
     );