+++ packages/swirl-components/src/components/swirl-time-input/swirl-time-input.stories.ts
@@ -1,0 +1,45 @@
+import { generateStoryElement } from "../../utils";
+import Docs from "./swirl-time-input.mdx";
+
+export default {
+  argTypes: {
+    format: {
+      description: "See https://date-fns.org/v2.29.3/docs/format.",
+      table: {
+        type: {
+          detail: '"h", "H", "m", "s", and separators (e.g. ":")',
+          summary: "Allowed patterns",
+        },
+      },
+    },
+    locale: {
+      control: { type: "object" },
+    },
+  },
+  component: "swirl-time-input",
+  tags: ["autodocs"],
+  parameters: {
+    docs: {
+      page: Docs,
+    },
+  },
+  title: "Components/SwirlTimeInput",
+};
+
+const Template = (args) => {
+  const formControl = document.createElement("swirl-form-control");
+  const element = generateStoryElement("swirl-time-input", args);
+
+  formControl.setAttribute("label", "Time");
+  formControl.append("\n  ", element, "\n");
+
+  element.addEventListener("valueChange", (event) => {
+    console.log(event);
+  });
+
+  return formControl;
+};
+
+export const SwirlTimeInput = Template.bind({});
+
+SwirlTimeInput.args = {};