+++ packages/swirl-components/stencil.config.ts
@@ -33,6 +33,7 @@ const angularValueAccessorBindings: ValueAccessorConfig[] = [
       "swirl-radio-group",
       "swirl-search",
       "swirl-text-input",
+      "swirl-time-input",
     ],
     event: "valueChange",
     targetAttr: "value",
@@ -100,6 +101,9 @@ export const config: Config = {
       valueAccessorConfigs: angularValueAccessorBindings,
     }),
   ],
+  extras: {
+    enableImportInjection: true,
+  },
   plugins: [
     postcss({
       plugins: [autoprefixer(), postcssNested(), postcssCustomMedia()],