+++ packages/swirl-tokens/dart/lib/styles.light.dart
@@ -4,7 +4,11 @@
 //
 
 // Do not edit directly
+<<<<<<< HEAD
 // Generated on Tue, 13 Jun 2023 07:57:16 GMT
+=======
+// Generated on Wed, 07 Jun 2023 12:55:03 GMT
+>>>>>>> main